Add Swift version to podspec file + format & lint the code
Cocoapods 1.4.0 includes the ability to specify the Swift version in the podspec (previously this could be done by creating a separate file in the repo). This allows Cocoapods to know which target to set for the Pod - instead of going for a default value, which can cause troubles when mixing Swift 4 and 3 code.

This change adds the Swift version to the Podspec.

The other changes includes some basic formatting and linting (mostly removal of whitespace and following Swift styleguides).

Also updates the Podspec version.